+++ code/compiler.samples/content-publishing/vite.config.mts
@@ -2,5 +2,6 @@ import { Config } from '../../../config.mjs';
 
 export default Config.vite(import.meta.url, (e) => {
   e.lib();
+  e.target('web');
   e.externalDependency(e.ctx.deps.map((d) => d.name));
 });
